# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple api that shows a weather forecast.",
        "title": "Weather forecast",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/api/cityList": {
            "get": {
                "description": "Get all cities avaliable for forecast",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "GetCityList",
                "operationId": "apiGetCityList",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/full_info/{cityName}/{date}": {
            "get": {
                "description": "Show full weather info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "FullInfo",
                "operationId": "fullInfo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "City name in any case",
                        "name": "cityName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Date in 2022-10-26T12:00:00Z format",
                        "name": "date",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/L3WB.FullCityGeoAndWeatherData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/shortInfo/{cityName}": {
            "get": {
                "description": "Show short weather info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "ShortCityInfo",
                "operationId": "shortInfo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "City name in any case",
                        "name": "cityName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/L3WB.ShortCityWeatherData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "L3WB.CityWeatherData": {
            "type": "object",
            "properties": {
                "clouds": {
                    "type": "object",
                    "properties": {
                        "all": {
                            "type": "integer"
                        }
                    }
                },
                "dt": {
                    "type": "integer"
                },
                "dt_txt": {
                    "type": "string"
                },
                "main": {
                    "type": "object",
                    "properties": {
                        "feels_like": {
                            "type": "number"
                        },
                        "grnd_level": {
                            "type": "integer"
                        },
                        "humidity": {
                            "type": "integer"
                        },
                        "pressure": {
                            "type": "integer"
                        },
                        "sea_level": {
                            "type": "integer"
                        },
                        "temp": {
                            "type": "number"
                        },
                        "temp_kf": {
                            "type": "number"
                        },
                        "temp_max": {
                            "type": "number"
                        },
                        "temp_min": {
                            "type": "number"
                        }
                    }
                },
                "pop": {
                    "type": "number"
                },
                "rain": {
                    "type": "object",
                    "properties": {
                        "3h": {
                            "type": "number"
                        }
                    }
                },
                "sys": {
                    "type": "object",
                    "properties": {
                        "pod": {
                            "type": "string"
                        }
                    }
                },
                "visibility": {
                    "type": "integer"
                },
                "weather": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "description": {
                                "type": "string"
                            },
                            "icon": {
                                "type": "string"
                            },
                            "id": {
                                "type": "integer"
                            },
                            "main": {
                                "type": "string"
                            }
                        }
                    }
                },
                "wind": {
                    "type": "object",
                    "properties": {
                        "deg": {
                            "type": "integer"
                        },
                        "gust": {
                            "type": "number"
                        },
                        "speed": {
                            "type": "number"
                        }
                    }
                }
            }
        },
        "L3WB.FullCityGeoAndWeatherData": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "object",
                    "properties": {
                        "coord": {
                            "type": "object",
                            "properties": {
                                "lat": {
                                    "type": "number"
                                },
                                "lon": {
                                    "type": "number"
                                }
                            }
                        },
                        "country": {
                            "type": "string"
                        },
                        "id": {
                            "type": "integer"
                        },
                        "name": {
                            "type": "string"
                        },
                        "population": {
                            "type": "integer"
                        },
                        "sunrise": {
                            "type": "integer"
                        },
                        "sunset": {
                            "type": "integer"
                        },
                        "timezone": {
                            "type": "integer"
                        }
                    }
                },
                "cityWeatherData": {
                    "$ref": "#/definitions/L3WB.CityWeatherData"
                }
            }
        },
        "L3WB.ShortCityWeatherData": {
            "type": "object",
            "properties": {
                "avgTemp": {
                    "type": "number"
                },
                "cityName": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "date": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "handler.errorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}